+++ aquadoggo/src/db/stores/mod.rs
@@ -4,6 +4,7 @@ pub mod document;
 mod entry;
 mod log;
 mod operation;
+mod schema;
 #[cfg(test)]
 pub mod test_utils;