+++ aquadoggo/src/db/traits/mod.rs
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 mod document;
+mod schema;
 
 pub use document::DocumentStore;
+pub use schema::SchemaStore;